# specs/002-fullstack-web-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Application API
  description: |
    RESTful API for Phase II Todo Application with JWT authentication.

    **Authentication**: All protected endpoints require a valid JWT access token in the Authorization header.

    **User Isolation**: All task operations are filtered by the authenticated user's ID.
  version: 1.0.0
  contact:
    name: Todo App Support
    email: support@todoapp.example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todoapp.example.com
    description: Production server (Phase V)
...
security:
  - BearerAuth: []
paths:
...
  /api/tasks:
    get:
      summary: List all tasks
      description: Retrieve all tasks for the authenticated user, sorted by completion status and creation date
      tags:
        - tasks
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input (empty title, title too long, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyTitle:
                  value:
                    detail: "Title cannot be empty"
                titleTooLong:
                  value:
                    detail: "Title too long (max 200 characters)"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}:
    put:
      summary: Update a task
      description: Update title and/or description of an existing task (authorization check enforced)
      tags:
        - tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"

    delete:
      summary: Delete a task
      description: Permanently delete a task (authorization check enforced)
      tags:
        - tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to delete
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/toggle:
    patch:
      summary: Toggle task completion
      description: Toggle task between completed and incomplete status (authorization check enforced)
      tags:
        - tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to toggle
      responses:
        '200':
          description: Task toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login (15 minute expiration)
  schemas:
...
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          example: "Buy groceries"
          description: Task title (max 200 characters)
        description:
          type: string
          maxLength: 2000
          example: "Milk, eggs, bread"
          description: Optional task description (max 2000 characters)
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          example: "Buy groceries and snacks"
          description: Updated task title (max 200 characters)
        description:
          type: string
          maxLength: 2000
          example: "Milk, eggs, bread, chips"
          description: Updated task description (max 2000 characters)
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Task ID
        title:
          type: string
          example: "Buy groceries"
          description: Task title
        description:
          type: string
          example: "Milk, eggs, bread"
          description: Task description
        completed:
          type: boolean
          example: false
          description: Completion status
        created_at:
          type: string
          format: date-time
          example: "2025-12-30T10:05:00Z"
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          example: "2025-12-30T10:05:00Z"
          description: Last modification timestamp
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Invalid credentials"
          description: Error message